+++ sdk/ai/azopenai/CHANGELOG.md
@@ -1,10 +1,13 @@
 # Release History
 
-## 0.2.1 (Unreleased)
+## 0.3.0 (Unreleased)
 
 ### Features Added
+- Support for Whisper audio APIs for transcription and translation using `GetAudioTranscription` and `GetAudioTranslation`.
 
 ### Breaking Changes
+- ChatChoiceContentFilterResults content filtering fields are now all typed as ContentFilterResult, instead of unique types for each field.
+- `PromptAnnotations` renamed to `PromptFilterResults` in `ChatCompletions` and `Completions`.
 
 ### Bugs Fixed